Always crate a ::Target
This is one more layer of (Chrome) indirection, but it allows me
to unify and simplify the code on the Perl side a lot. Or rather,
it allows me to keep the code almost as simple as it is now,
instead of catering for both cases of direct and indirect connection.

Also, it makes handling of multiple tabs/windows possible in the first
place, and maybe even discovery of other network-local browsers.